We drink to forget the coming storm
This project  nearing its end. I still remember the plan I had for my game. there were going to be real characters, not just a block. (well you would control a block as tome point in the game bet that's not important) I did a lot of world building without making the world and that's where I fell I made my biggest mistakes. I make that mistake a lot when I dm also so this will be something I need to work on next semester. Even with this semester reaching its end I still want to keep working on this project in  the future. Ill make another post on this tomorrow or when I can.  besides that the game right now in its current form is like a boring cake. I have all the ingredients and the pots and pans, now I just need to put it together. But I wont be baking it. Not yet.